add a command to list names of clusters
Adds a `--list` or `-l` optional argument which can list the names of
the clusters in the config file.

This might be useful if you are the type of person like me who often
forgets the name of the clusters in his config file and gets tired of
opening the config file up.